refactor: move files around and renamings
