+++ src/components/Inputs/index.tsx
@@ -1,6 +1,7 @@
-export { TextInput } from './TextInput'
-export { SelectInput } from './SelectInput'
-export { TextAreaInput } from './TextAreaInput'
 export { NumberInput } from './NumberInput'
 export { MillisecondInput } from './MillisecondInput'
+export { SelectInput } from './SelectInput'
+export { TextInput } from './TextInput'
+export { TextAreaInput } from './TextAreaInput'
 export { Toggle } from './Toggle'
+export { URLInput } from './URLInput'
