--- src/pages/Run/components/RunConfigurationPanel/components/RunConfigurationPanelHeaders/core/RunConfigurationPanelHeaders.typings.tsx
@@ -1,4 +1,0 @@
-export interface IProps {
-  headers: { key: string; values: string[] }[]
-  setHeaders: (headers: { key: string; values: string[] }[]) => void
-}
