+++ src/pages/Run/components/RunConfigurationPanel/components/index.ts
@@ -1,0 +1,3 @@
+export { MethodSelect } from './MethodSelect'
+export { HeadersForm } from './HeadersForm'
+export { TestsForm } from './TestsForm'
