--- src/pages/Run/components/RunConfigurationPanel/components/index.tsx
@@ -1,4 +1,0 @@
-export { MethodInput } from './MethodInput'
-export { RunConfigurationPanelHeaders } from './RunConfigurationPanelHeaders'
-export { RunConfigurationPanelTests } from './RunConfigurationPanelTests'
-export { URLInput } from './URLInput'
